killed

+++ Fitnow_structure/app/Http/Controllers/ProgressController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Progress;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProgressController extends Controller
 {
@@ -27,6 +28,17 @@ class ProgressController extends Controller
         //
     }
 
+    public function showUserProgress()
+    {
+        $userID = Auth::id();
+        $progress = Progress::where('userID', $userID)->get();
+
+        $data = [
+            'progress' => $progress,
+        ];
+        return response()->json($data, 200);
+    }
+
     /**
      * Store a newly created resource in storage.
      */
@@ -41,6 +53,8 @@ class ProgressController extends Controller
             'calve_thickness' => 'required',
         ]);
 
+        $userID = Auth::id();
+
         $success = Progress::create([
             'weight' => $validatedData['weight'],
             'height' => $validatedData['height'],
@@ -48,15 +62,16 @@ class ProgressController extends Controller
             'bicep_thickness' => $validatedData['bicep_thickness'],
             'pec_width' => $validatedData['pec_width'],
             'calve_thickness' => $validatedData['calve_thickness'],
+            'userID' => $userID,
         ]);
 
-        if($success){
+        if ($success) {
             $data = [
-                'message'=>'progress added succefully!'
+                'message' => 'progress added succefully!'
             ];
             return response()->json($data, 200);
-        }else{
-            return response()->json(['message'=>'unexpected error'], 422);
+        } else {
+            return response()->json(['message' => 'unexpected error'], 422);
         }
     }
 
@@ -81,29 +96,37 @@ class ProgressController extends Controller
      */
     public function update(Request $request, progress $progress)
     {
-        $validatedData = $request->validate([
-            'weight' => 'required',
-            'height' => 'required',
-            'waist_line' => 'required',
-            'bicep_thickness' => 'required',
-            'pec_width' => 'required',
-            'calve_thickness' => 'required',
-        ]);
+        $userID = Auth::id();
+        if ($userID == $progress->userID) {
+            $validatedData = $request->validate([
+                'weight' => 'required',
+                'height' => 'required',
+                'waist_line' => 'required',
+                'bicep_thickness' => 'required',
+                'pec_width' => 'required',
+                'calve_thickness' => 'required',
+            ]);
 
-        $success = $progress->update([
-            'weight' => $validatedData['weight'],
-            'height' => $validatedData['height'],
-            'waist_line' => $validatedData['waist_line'],
-            'bicep_thickness' => $validatedData['bicep_thickness'],
-            'pec_width' => $validatedData['pec_width'],
-            'calve_thickness' => $validatedData['calve_thickness'],
-        ]);
+            $success = $progress->update([
+                'weight' => $validatedData['weight'],
+                'height' => $validatedData['height'],
+                'waist_line' => $validatedData['waist_line'],
+                'bicep_thickness' => $validatedData['bicep_thickness'],
+                'pec_width' => $validatedData['pec_width'],
+                'calve_thickness' => $validatedData['calve_thickness'],
+            ]);
 
-        if($success){
-            $data = [
-                'message'=>'progress updated succefully!'
-            ];
-            return response()->json($data, 200);
+            if ($success) {
+                $data = [
+                    'message' => 'progress updated succefully!'
+                ];
+                return response()->json($data, 200);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Unauthorized',
+            ], 401);
         }
     }
 
@@ -112,30 +135,46 @@ class ProgressController extends Controller
      */
     public function destroy(progress $progress)
     {
-        $progress->delete();
+        $userID = Auth::id();
+        if ($userID == $progress->userID) {
 
-        $data = [
-            'message'=>'progress deleted succefully!'
-        ];      
-        return response()->json($data, 200);
+            $progress->delete();
+
+            $data = [
+                'message' => 'progress deleted succefully!'
+            ];
+            return response()->json($data, 200);
+        }else{
+            return response()->json([
+                'status' => false,
+                'message' => 'Unauthorized',
+            ], 401);
+        }
     }
 
     public function updateStatus(Request $request, progress $progress)
     {
-        $validatedData = $request->validate([
-            'status' => 'required',
-        ]);
+        $userID = Auth::id();
+        if ($userID == $progress->userID) {
+            $validatedData = $request->validate([
+                'status' => 'required',
+            ]);
 
-        $success = $progress->update([
-            'status' => $validatedData['status'],
-        ]);
+            $success = $progress->update([
+                'status' => $validatedData['status'],
+            ]);
 
-        if($success){
-            $data = [
-                'message'=>'status changed succefully!'
-            ];
-            return response()->json($data, 200);
+            if ($success) {
+                $data = [
+                    'message' => 'status changed succefully!'
+                ];
+                return response()->json($data, 200);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Unauthorized',
+            ], 401);
         }
     }
-
 }
